tpm2_createpolicy: move build_policy() function to tpm2 policy library
The build_policy() function logic is quite generic and so it can be used
by other tools. For example the tpm2_unseal tool also needs to build PCR
policies, so it can reuse this function instead of open coding it again.
